[game_skyrimse] Fixed Image colours, most of the save-time, and the plugin listing
Figured out a few of the changes to the savefile, Apparently the
filetime that the program reads is innacurate and about 6 hours ahead.
It's now mostly correct.
The Image problem occured because there was just nonsense directly
before the image started, but after the dimensions.
The plugins seem to start a variable number of bytes after the image
ends, but I used the presence of two 0 bytes in a row to determine the
location in the save to start reading the plugins.