[game_gamebryo] First Attempt at reading plugins with lz4 compression
Since no one has seen a savefile with zlib compression, I am ignoring
that until someone does, so I can actually view the save file to
determine if it works.
I'm not using the stream functionality of lz4 decompression currently
due to that being rather confusing. It also isn't too slow to do it this
way.