[modorganizer] Fix matching highlights for mod and plugin selections
The selectionChanged event used before requires subtractive/additive logic to work correctly.  A separate event is fired for each row and each column for each row that is selected or deselected.  The previous logic assumed all of the selection was passed in at once.

The logic has been changed to recalculate all highlights each time the event is fired.  This is poorly optimized in regards to performance but makes the function work.