[modorganizer] Fix for Steam being run as administrator
Previously, when Steam was run as administrator, MO couldn't
access the process therefore couldn't detect it.  Now, an
alternative method for listing processes is now being used
in order to determine if Steam is running.

Additionally, MO will detect if Steam is admin and MO is not
and prompt the user if they want to restart MO as admin.