[modorganizer] Fix md5 hashing for large download files
Previously, the md5 hashing was broken for download files larger
than about 2GB.  This was due to the maximum size of a QByteArray.
The hashing was broken up into 10MB chunks to correct this issue.

Additionally, a progress bar was added as hashing large files would
lock up the GUI.