[modorganizer] Improve selection of mod ID when querying info
The mod ID selection was fairly unreliable when the mod author decided
to use certain values in the mod name or filename.  A filename of
"K-9 - Armor-1-2-3" would automatically set the mod ID to 9 with no
chance for the user to correct it.  Other filenames would only present
2 choices of mod ID to the user, neither being correct.

The primary intent of this change is to present the user with all the
numbers found in the filename as options for the mod ID.  This should
allow the correct mod ID to be easily chosen in all cases.