[installer_omod] Prevent nested UpdateProgressValueSlot calls by using a blocking connection
Becuase setProgress calls processEvents, it was possible for the next
UpdateProgressValue call to signal while the slot was still running,
causing unwanted recursion and a StackOverflowException which crashes
the whole application without logging anything.

Switching to a blocking connection makes extraction wait until the
function returns, solving the problem.

To make that work nicely (so things only block when needed) keeping
track of the percentage was moved to CodeProgress so that the singal
only gets emitted when something genuinely needs running on the GUI
thread.