[basic_games] Pathlib stuff
`os.path` is old and less safe than Pathlib, so I removed it while I wanted to meddle with this file anyway.

It looked suspicious that `getCreationTime` was returning the modification time instead of the creation time, so I changed that from `mtime` to `ctime`.

I also added handling for `value` being a `Path` object as that's the type for the thing it represents. I might not have needed the explicit `str` call as Path is implicitly convertible to `str` lots of the time, but I've not actually tested the change so wanted to play it safe.