[modorganizer] Add error checking to lockedorder.txt processing
Errors now checked:
1. Plugin listed in lockedorder.txt does not exist
2. Priority in lockedorder.txt is negative
3. Priority in lockedorder.txt is greater than the total number of plugins

Note that lockedorder.txt is not rewritten after these errors occur as
that seems most likely to be the desired behavior...?  That is, preserving
the state of lockedorder.txt is more important than anything making sense.
Though negative priorities don't make sense and never will so just get rid
of that if the lockedorder.txt is written to at some point.