[modorganizer] Fix forced load libraries
This fixes a couple issues.  First, the default libraries forced by
the game plugins were not being loaded.  Second, when the checkbox
was disabled, the forced load settings would be wiped out.

Both were fixed by always loading and saving the state of the
forced loads regardless of the global enabled state.