[basic_games] Add support for Divinity Original Sin
This includes the original (untested, mostly just for Nexus
download support) and the enhanced edition (actually somewhat
tested).

This is some basic support for mods installed in the game
folder.  There's also a location for actual packaged mods
in the Documents folder that I hope to add support for at
some point.