[basic_games] Add Mass Effect Legendary Edition plugin
All the weird stuff here is done to get around the Origin DRM. The
only way I've found to make MO2 and Origin play nicely together is
to run Origin through MO2. Therefore, the plugin makes sure Origin
is dead when running an executable.

Then because Origin is still alive and hooked by MO2 (probably bad),
it tries to kill Origin after the game is closed. However, the
onFinishedRun callback is only called when all hooked processes are
closed so a a thread is used to monitor the processes.

Psutil is included from PyPi for accessing process data.